refactor: split item_service into focused modules (DEBT-1 D1-B)
item_service.py 2,403 → 1,502 lines (core CRUD/lookup/listing, keeps the
full public surface via re-exports). Extracted verbatim:
- item_cost_service.py (225) — cost rollup/recost
- item_duplicate_service.py (255) — duplicate detection/merge
- item_import_service.py (517) — bulk import/CSV/preview

Mechanical move only; no logic changes. Proof: full backend suite
(5,489 passed, 13 skipped) green with zero test modifications.

Agent-Session: claude-filaops-pm-20260610-uibatch
