+++ backend/app/services/item_service.py
@@ -2,23 +2,25 @@
 Item Service — CRUD and business logic for items and categories.
 
 Items extends Products with inventory info, categories, reorder logic,
-CSV import, and cost recalculation. Extracted from items.py (ARCHITECT-003).
+and pricing. Extracted from items.py (ARCHITECT-003).
+
+DEBT-1 D1-B split: cost recalculation lives in item_cost_service,
+duplication in item_duplicate_service, CSV import in item_import_service.
+All historical public names remain importable from this module via the
+re-export block at the bottom.
 """
-import csv
-import io
 from datetime import datetime, timezone
 from decimal import Decimal
 
 from fastapi import HTTPException
-from sqlalchemy import or_, func, desc
+from sqlalchemy import or_, func
 from sqlalchemy.orm import Session, joinedload
 
 from app.logging_config import get_logger
 from app.models import Product, ItemCategory, Inventory, BOM, BOMLine
-from app.models.manufacturing import Routing, RoutingOperation, RoutingOperationMaterial
+from app.models.manufacturing import Routing
 from app.core.utils import get_or_404, check_unique_or_400
 from app.core.uom_config import DEFAULT_MATERIAL_UOM, get_default_material_sku_prefix
-from app.services.bom_management_service import recalculate_bom_cost
 
 logger = get_logger(__name__)
 
@@ -1352,206 +1354,6 @@ def bulk_update_items(
     }
 
 
-def calculate_item_cost(item: Product, db: Session) -> dict:
-    """
-    Calculate standard cost for an item.
-
-    Returns a dict with: bom_id, bom_cost, routing_id, routing_cost,
-    purchase_cost, total_cost, cost_source.
-    """
-    bom_cost = 0.0
-    routing_cost = 0.0
-    purchase_cost = 0.0
-    bom_id = None
-    routing_id = None
-    cost_source = None
-
-    bom = (
-        db.query(BOM)
-        .filter(BOM.product_id == item.id, BOM.active.is_(True))
-        .first()
-    )
-
-    # Check for active routing (independent of BOM — variant items may
-    # have a routing with materials but no BOM)
-    routing = (
-        db.query(Routing)
-        .options(
-            joinedload(Routing.operations)
-            .joinedload(RoutingOperation.work_center),
-            joinedload(Routing.operations)
-            .joinedload(RoutingOperation.materials)
-            .joinedload(RoutingOperationMaterial.component),
-        )
-        .filter(Routing.product_id == item.id, Routing.is_active.is_(True))
-        .order_by(desc(Routing.version))
-        .first()
-    )
-
-    if bom or routing:
-        cost_source = "manufactured"
-
-        if bom:
-            bom_id = bom.id
-
-        # Collect component IDs costed via routing operations to avoid
-        # double-counting the same material in both BOM and routing.
-        # Design: if a component_id is on routing, ALL BOM lines with
-        # that component are excluded (routing owns that material fully).
-        # This matches override_map behavior in duplicate_item().
-        routing_material_ids = set()
-        if routing:
-            routing.recalculate_totals()
-            routing_cost = float(routing.total_cost) if routing.total_cost else 0.0
-            routing_id = routing.id
-            for op in routing.operations:
-                if op.is_active:
-                    for mat in op.materials:
-                        # Only include per-unit materials — batch/order are
-                        # excluded from material_cost so must not exclude
-                        # the matching BOM line either (would lose the cost)
-                        is_per_unit = not mat.quantity_per or str(mat.quantity_per).strip().lower() == "unit"
-                        if is_per_unit and mat.extended_cost and mat.extended_cost > 0:
-                            routing_material_ids.add(mat.component_id)
-
-        # BOM cost calculation (only if BOM exists)
-        if bom:
-            # Always store full BOM cost (what the BOM page shows)
-            full_bom_cost = recalculate_bom_cost(bom, db)
-            bom.total_cost = full_bom_cost
-
-            # For item STD cost, subtract routing-owned material costs from
-            # the full BOM to avoid double-counting. Uses only_component_ids
-            # to compute just the overlap — components are already in the
-            # SQLAlchemy session cache from the first call.
-            if routing_material_ids:
-                overlap_cost = recalculate_bom_cost(
-                    bom, db, only_component_ids=routing_material_ids
-                )
-                bom_cost = float(full_bom_cost - overlap_cost)
-            else:
-                bom_cost = float(full_bom_cost)
-
-        total_cost = bom_cost + routing_cost
-    else:
-        cost_source = "purchased"
-        if item.standard_cost and item.standard_cost > 0:
-            purchase_cost = float(item.standard_cost)
-        elif item.average_cost:
-            purchase_cost = float(item.average_cost)
-        elif item.last_cost:
-            purchase_cost = float(item.last_cost)
-
-        total_cost = purchase_cost
-
-    return {
-        "bom_id": bom_id,
-        "bom_cost": bom_cost,
-        "routing_id": routing_id,
-        "routing_cost": routing_cost,
-        "purchase_cost": purchase_cost,
-        "total_cost": total_cost,
-        "cost_source": cost_source,
-    }
-
-
-def recost_item(db: Session, item_id: int) -> dict:
-    """Recost a single item and return the result."""
-    item = get_item(db, item_id)
-    cost_data = calculate_item_cost(item, db)
-
-    old_cost = float(item.standard_cost) if item.standard_cost else 0
-    item.standard_cost = cost_data["total_cost"]
-    item.updated_at = datetime.now(timezone.utc)
-    db.commit()
-    db.refresh(item)
-
-    logger.info(
-        f"Recost item {item.sku}: ${old_cost:.4f} -> ${cost_data['total_cost']:.4f}"
-    )
-
-    return {
-        "id": item.id,
-        "sku": item.sku,
-        "name": item.name,
-        "old_cost": old_cost,
-        "new_cost": cost_data["total_cost"],
-        "cost_source": cost_data["cost_source"],
-        "bom_id": cost_data["bom_id"],
-        "bom_cost": cost_data["bom_cost"],
-        "routing_id": cost_data["routing_id"],
-        "routing_cost": cost_data["routing_cost"],
-        "purchase_cost": cost_data["purchase_cost"],
-        "message": f"Standard cost updated: ${old_cost:.4f} -> ${cost_data['total_cost']:.4f}",
-    }
-
-
-def recost_all_items(
-    db: Session,
-    *,
-    item_type: str | None = None,
-    category_id: int | None = None,
-    cost_source_filter: str | None = None,
-) -> dict:
-    """
-    Recost all items matching filters.
-
-    Returns dict with: updated, skipped, items (list of results).
-    """
-    query = db.query(Product).filter(Product.active.is_(True))
-
-    if item_type:
-        query = query.filter(Product.item_type == item_type)
-
-    if category_id:
-        category_ids = get_category_and_descendants(db, category_id)
-        query = query.filter(Product.category_id.in_(category_ids))
-
-    items = query.all()
-
-    updated = 0
-    skipped = 0
-    results = []
-
-    for item in items:
-        cost_data = calculate_item_cost(item, db)
-
-        if cost_source_filter and cost_data["cost_source"] != cost_source_filter:
-            continue
-
-        if cost_data["total_cost"] == 0:
-            skipped += 1
-            continue
-
-        old_cost = float(item.standard_cost) if item.standard_cost else 0
-        item.standard_cost = cost_data["total_cost"]
-        item.updated_at = datetime.now(timezone.utc)
-        updated += 1
-
-        results.append(
-            {
-                "id": item.id,
-                "sku": item.sku,
-                "old_cost": old_cost,
-                "new_cost": cost_data["total_cost"],
-                "cost_source": cost_data["cost_source"],
-                "bom_cost": cost_data["bom_cost"],
-                "routing_cost": cost_data["routing_cost"],
-                "purchase_cost": cost_data["purchase_cost"],
-            }
-        )
-
-    db.commit()
-
-    logger.info(f"Recost all: {updated} items updated, {skipped} skipped")
-
-    return {
-        "updated": updated,
-        "skipped": skipped,
-        "items": results,
-    }
-
-
 # ---------------------------------------------------------------------------
 # Suggest Prices
 # ---------------------------------------------------------------------------
@@ -1663,741 +1465,38 @@ def apply_suggested_prices(
 
 
 # ---------------------------------------------------------------------------
-# CSV Import Column Mappings
+# Backward-compatibility re-exports (DEBT-1 D1-B mechanical split)
+#
+# item_service keeps its full historical public surface; new code should
+# import from the focused modules directly. Placed at the bottom so the
+# focused modules can resolve item_service helpers at call time.
 # ---------------------------------------------------------------------------
-
-# Marketplace column mappings for SKU
-_SKU_COLUMNS = [
-    "sku", "SKU", "Sku", "product_sku", "Product SKU", "product-sku",
-    "Variant SKU", "variant_sku", "variant-sku", "VariantSKU",
-    "SKU Code", "sku_code", "sku-code", "SKUCode",
-    "ASIN", "asin", "Amazon ASIN",
-    "Product Code", "product_code", "product-code",
-    "Item SKU", "item_sku", "item-sku",
-    "Product ID", "product_id", "product-id",
-]
-
-# Marketplace column mappings for Name
-_NAME_COLUMNS = [
-    "name", "Name", "product_name", "Product Name", "product-name",
-    "title", "Title", "Product Title", "product-title",
-    "Variant Title", "variant_title", "variant-title", "VariantTitle",
-    "Product Title", "product_title", "product-title", "ProductTitle",
-    "Item Name", "item_name", "item-name",
-]
-
-# Marketplace column mappings for Description
-_DESCRIPTION_COLUMNS = [
-    "description", "Description", "product_description", "Product Description",
-    "Body (HTML)", "body_html", "Body", "body", "Body HTML",
-    "Short Description", "short_description", "short-description", "Short description",
-    "Long Description", "long_description", "long-description",
-    "Product Description", "product-description",
-    "Item Description", "item_description",
-]
-
-# Marketplace column mappings for Price
-_PRICE_COLUMNS = [
-    "selling_price", "Selling Price", "selling-price",
-    "price", "Price",
-    "Variant Price", "variant_price", "variant-price", "VariantPrice",
-    "Variant Compare At Price", "variant_compare_at_price", "variant-compare-at-price",
-    "Sale price", "sale_price", "sale-price", "Sale Price",
-    "Regular price", "regular_price", "regular-price", "Regular Price",
-    "Unit Price", "unit_price", "unit-price", "UnitPrice",
-    "Retail Price", "retail_price", "retail-price",
-    "List Price", "list_price", "list-price",
-]
-
-# Marketplace column mappings for Cost
-_COST_COLUMNS = [
-    "standard_cost", "Standard Cost", "standard-cost",
-    "cost", "Cost",
-    "Variant Cost", "variant_cost", "variant-cost", "VariantCost",
-    "Wholesale Price", "wholesale_price", "wholesale-price",
-    "Cost Price", "cost_price", "cost-price", "CostPrice",
-    "Purchase PPU", "purchase_ppu", "purchase-ppu",
-    "Item Subtotal", "item_subtotal",
-    "Purchase Cost", "purchase_cost", "purchase-cost",
-    "Unit Cost", "unit_cost", "unit-cost",
-    "Wholesale Cost", "wholesale_cost",
-]
-
-# UOM column mappings
-_UNIT_COLUMNS = [
-    "unit", "Unit", "UOM", "uom", "Unit of Measure", "unit_of_measure",
-]
-
-_PURCHASE_UOM_COLUMNS = [
-    "purchase_uom", "Purchase UOM", "purchase_unit", "Purchase Unit",
-    "buying_unit", "Buying Unit", "order_unit", "Order Unit",
-]
-
-
-def _get_csv_column_value(row: dict, possible_names: list[str]) -> str:
-    """Get value from CSV row using case-insensitive column name matching."""
-    for col in possible_names:
-        if row.get(col, "").strip():
-            return row.get(col, "").strip()
-    return ""
-
-
-def _parse_price(price_str: str) -> float | None:
-    """Parse price string, removing currency symbols."""
-    if not price_str:
-        return None
-    price_clean = (
-        price_str.replace("$", "")
-        .replace(",", "")
-        .replace("€", "")
-        .replace("£", "")
-        .strip()
-    )
-    try:
-        return float(price_clean)
-    except ValueError:
-        return None
-
-
-def _strip_html(text: str) -> str:
-    """Strip HTML tags from text."""
-    import re
-
-    if "<" in text and ">" in text:
-        return re.sub(r"<[^>]+>", "", text).strip()
-    return text
-
-
-def _normalize_import_item_type(value: str | None, default: str) -> str:
-    """Normalize item type names from CSV exports and shop-owner shorthand."""
-    item_type_raw = (value or "").strip().lower()
-    if not item_type_raw:
-        return default
-
-    item_type_map = {
-        "simple": "finished_good",
-        "variable": "finished_good",
-        "finished_good": "finished_good",
-        "component": "component",
-        "packaging": "packaging",
-        "box": "packaging",
-        "mailers": "packaging",
-        "mailer": "packaging",
-        "supply": "supply",
-        "service": "service",
-        "material": "material",
-        "filament": "material",
-        "raw_material": "material",
-    }
-    return item_type_map.get(item_type_raw, default)
-
-
-def import_items_from_csv(
-    db: Session,
-    *,
-    file_content: bytes,
-    update_existing: bool = False,
-    default_item_type: str = "finished_good",
-    default_category_id: int | None = None,
-) -> dict:
-    """
-    Import items from CSV file content.
-
-    Returns dict with keys: total_rows, created, updated, skipped, errors, warnings.
-    """
-    from app.services.product_uom_service import get_recommended_uoms, validate_product_uoms
-
-    try:
-        text = file_content.decode("utf-8")
-    except UnicodeDecodeError:
-        text = file_content.decode("latin-1")
-
-    if text.startswith("\ufeff"):
-        text = text[1:]
-
-    reader = csv.DictReader(io.StringIO(text))
-
-    result = {
-        "total_rows": 0,
-        "created": 0,
-        "updated": 0,
-        "skipped": 0,
-        "errors": [],
-        "warnings": [],
-    }
-
-    for row_num, row in enumerate(reader, start=2):
-        result["total_rows"] += 1
-
-        try:
-            # Find SKU
-            sku = _get_csv_column_value(row, _SKU_COLUMNS).upper()
-            if not sku:
-                result["errors"].append({
-                    "row": row_num,
-                    "error": "SKU is required",
-                })
-                result["skipped"] += 1
-                continue
-
-            # Find name
-            name = _get_csv_column_value(row, _NAME_COLUMNS)
-            if not name:
-                result["errors"].append({
-                    "row": row_num,
-                    "error": "Name is required",
-                    "sku": sku,
-                })
-                result["skipped"] += 1
-                continue
-
-            # Check if exists
-            existing = db.query(Product).filter(Product.sku == sku).first()
-
-            if existing:
-                # Protect seeded example items
-                if existing.sku.startswith("SEED-EXAMPLE-"):
-                    result["errors"].append({
-                        "row": row_num,
-                        "error": f"SKU '{sku}' is a seeded example item",
-                        "sku": sku,
-                    })
-                    result["skipped"] += 1
-                    continue
-
-                if not update_existing:
-                    result["skipped"] += 1
-                    continue
-
-                # Update existing
-                existing.name = name
-
-                # Update description
-                desc = _get_csv_column_value(row, _DESCRIPTION_COLUMNS)
-                if desc:
-                    existing.description = _strip_html(desc)
-
-                # Update unit
-                unit_value = _get_csv_column_value(row, _UNIT_COLUMNS).upper()
-                if unit_value:
-                    existing.unit = unit_value
-
-                # Update purchase_uom
-                purchase_uom_value = _get_csv_column_value(
-                    row, _PURCHASE_UOM_COLUMNS
-                ).upper()
-                if purchase_uom_value:
-                    existing.purchase_uom = purchase_uom_value
-
-                # Update item type
-                item_type_raw = (
-                    row.get("item_type", "")
-                    or row.get("Item Type", "")
-                    or row.get("Type", "")
-                ).strip()
-                if item_type_raw:
-                    existing.item_type = _normalize_import_item_type(
-                        item_type_raw, existing.item_type
-                    )
-
-                # Update category
-                _update_category_from_row(db, existing, row, default_category_id)
-
-                # Update cost
-                cost_str = _get_csv_column_value(row, _COST_COLUMNS)
-                cost = _parse_price(cost_str)
-                if cost is not None:
-                    existing.standard_cost = cost
-
-                # Update price
-                price_str = _get_best_price_from_row(row)
-                price = _parse_price(price_str)
-                if price is not None:
-                    existing.selling_price = price
-
-                # Update reorder point
-                if row.get("reorder_point"):
-                    try:
-                        existing.reorder_point = float(row["reorder_point"])
-                    except ValueError:
-                        pass
-
-                # Update UPC
-                upc = _get_upc_from_row(row)
-                if upc:
-                    existing.upc = upc
-
-                existing.updated_at = datetime.now(timezone.utc)
-
-                # Validate UOM configuration
-                is_valid, warning_msg = validate_product_uoms(db, existing)
-                if not is_valid:
-                    result["warnings"].append({
-                        "row": row_num,
-                        "sku": sku,
-                        "warning": warning_msg,
-                    })
-
-                result["updated"] += 1
-
-            else:
-                # Create new item
-                desc = _get_csv_column_value(row, _DESCRIPTION_COLUMNS)
-                description = _strip_html(desc) if desc else None
-
-                price_str = _get_best_price_from_row(row)
-                selling_price = _parse_price(price_str)
-
-                cost_str = _get_csv_column_value(row, _COST_COLUMNS)
-                standard_cost = _parse_price(cost_str)
-
-                # Handle category
-                final_category_id = _get_category_id_from_row(
-                    db, row, default_category_id
-                )
-
-                # Get item type
-                item_type_str = (
-                    row.get("item_type", "")
-                    or row.get("Item Type", "")
-                    or row.get("Type", "")
-                    or ""
-                ).strip() or default_item_type
-                item_type_str = _normalize_import_item_type(
-                    item_type_str, default_item_type
-                )
-
-                # Get unit from CSV
-                unit_value = _get_csv_column_value(row, _UNIT_COLUMNS).upper()
-                purchase_uom_value = _get_csv_column_value(
-                    row, _PURCHASE_UOM_COLUMNS
-                ).upper()
-                purchase_factor = None
-                is_raw_material = False
-
-                # Auto-detect UOMs based on item type, SKU, and category if not provided.
-                if not purchase_uom_value or not unit_value:
-                    (
-                        recommended_purchase,
-                        recommended_unit,
-                        is_raw_material,
-                        purchase_factor,
-                    ) = get_recommended_uoms(
-                        db,
-                        sku=sku,
-                        category_id=final_category_id,
-                        item_type=item_type_str,
-                    )
-                    if not purchase_uom_value:
-                        purchase_uom_value = recommended_purchase
-                    if not unit_value and is_raw_material:
-                        unit_value = recommended_unit
-
-                final_unit = unit_value or "EA"
-                final_purchase_uom = purchase_uom_value or final_unit
-                if purchase_factor is None:
-                    purchase_factor = 1
-
-                # Get reorder point
-                reorder_point = None
-                if row.get("reorder_point"):
-                    try:
-                        reorder_point = float(row["reorder_point"])
-                    except ValueError:
-                        pass
-
-                # Get UPC
-                upc = _get_upc_from_row(row)
-
-                item = Product(
-                    sku=sku,
-                    name=name,
-                    description=description,
-                    unit=final_unit,
-                    purchase_uom=final_purchase_uom,
-                    item_type=item_type_str,
-                    category_id=final_category_id,
-                    standard_cost=standard_cost,
-                    selling_price=selling_price,
-                    purchase_factor=purchase_factor,
-                    is_raw_material=is_raw_material,
-                    reorder_point=reorder_point,
-                    upc=upc,
-                    active=True,
-                )
-                db.add(item)
-                db.flush()
-
-                # Validate UOM configuration
-                is_valid, warning_msg = validate_product_uoms(db, item)
-                if not is_valid:
-                    result["warnings"].append({
-                        "row": row_num,
-                        "sku": sku,
-                        "warning": warning_msg,
-                    })
-
-                result["created"] += 1
-
-        except Exception as e:
-            result["errors"].append({
-                "row": row_num,
-                "error": str(e),
-                "sku": row.get("sku", ""),
-            })
-            result["skipped"] += 1
-
-    db.commit()
-
-    logger.info(
-        f"CSV import complete: {result['created']} created, "
-        f"{result['updated']} updated, {result['skipped']} skipped, "
-        f"{len(result['warnings'])} UOM warnings"
-    )
-
-    return result
-
-
-def _get_best_price_from_row(row: dict) -> str:
-    """Get the best price from CSV row, preferring sale price."""
-    for col in _PRICE_COLUMNS:
-        value = row.get(col, "").strip()
-        if value:
-            # Prefer sale price if available (WooCommerce)
-            if "sale" in col.lower():
-                return value
-    # Return first found as fallback
-    return _get_csv_column_value(row, _PRICE_COLUMNS)
-
-
-def _get_upc_from_row(row: dict) -> str | None:
-    """Get UPC/barcode from CSV row."""
-    upc_cols = [
-        "upc", "UPC", "barcode", "Barcode", "EAN", "GTIN",
-        "Product Code", "product_code", "ASIN", "asin",
-    ]
-    value = _get_csv_column_value(row, upc_cols)
-    return value if value else None
-
-
-def _update_category_from_row(
-    db: Session, product: Product, row: dict, default_category_id: int | None
-) -> None:
-    """Update product category from CSV row data."""
-    # Try category_id first (numeric)
-    category_id_raw = (
-        row.get("category_id", "")
-        or row.get("Category ID", "")
-        or row.get("category-id", "")
-    ).strip()
-    if category_id_raw:
-        try:
-            product.category_id = int(category_id_raw)
-            return
-        except ValueError:
-            pass
-
-    # Try category name
-    category_name_raw = (
-        row.get("Category", "")
-        or row.get("category", "")
-        or row.get("Categories", "")
-        or row.get("Product Category", "")
-        or row.get("Type", "")
-        or row.get("Product Type", "")
-    ).strip()
-
-    if category_name_raw:
-        # Handle WooCommerce comma-separated categories (take first)
-        if "," in category_name_raw:
-            category_name_raw = category_name_raw.split(",")[0].strip()
-
-        # Try to find category by name
-        category = (
-            db.query(ItemCategory)
-            .filter(ItemCategory.name.ilike(f"%{category_name_raw}%"))
-            .first()
-        )
-        if category:
-            product.category_id = category.id
-
-
-def _get_category_id_from_row(
-    db: Session, row: dict, default_category_id: int | None
-) -> int | None:
-    """Get category ID from CSV row data."""
-    # Try category_id first (numeric)
-    category_id_raw = (
-        row.get("category_id", "")
-        or row.get("Category ID", "")
-        or row.get("category-id", "")
-    ).strip()
-    if category_id_raw:
-        try:
-            return int(category_id_raw)
-        except ValueError:
-            pass
-
-    # Try category name
-    category_name_raw = (
-        row.get("Category", "")
-        or row.get("category", "")
-        or row.get("Categories", "")
-        or row.get("Product Category", "")
-        or row.get("Type", "")
-        or row.get("Product Type", "")
-    ).strip()
-
-    if category_name_raw:
-        if "," in category_name_raw:
-            category_name_raw = category_name_raw.split(",")[0].strip()
-
-        category = (
-            db.query(ItemCategory)
-            .filter(ItemCategory.name.ilike(f"%{category_name_raw}%"))
-            .first()
-        )
-        if category:
-            return category.id
-
-    return default_category_id
-
-
-# --- Duplicate Item ---
-
-def duplicate_item(
-    db: Session,
-    source_item_id: int,
-    *,
-    new_sku: str,
-    new_name: str,
-    bom_line_overrides: list[dict] | None = None,
-) -> dict:
-    """
-    Duplicate a product: clone all fields with a new SKU/name,
-    copy the active BOM (if any), and apply component overrides.
-
-    Returns dict with: id, sku, name, has_bom, bom_id, message
-    """
-    source = get_item(db, source_item_id)
-
-    # Validate new SKU uniqueness
-    new_sku_upper = new_sku.upper().strip()
-    if not new_sku_upper:
-        raise HTTPException(status_code=400, detail="SKU cannot be blank")
-    new_name_clean = new_name.strip()
-    if not new_name_clean:
-        raise HTTPException(status_code=400, detail="Name cannot be blank")
-    check_unique_or_400(db, Product, "sku", new_sku_upper)
-
-    # Fields to exclude from copy:
-    # - Identity: id, sku, name, timestamps
-    # - External IDs: woocommerce, squarespace, legacy_sku, upc (unique per item)
-    # - Purchase history: average_cost, last_cost (no history for new item)
-    # - Per-variant assets: gcode_file_path, image_url (different per color/variant)
-    # - B2B restriction: customer_id (new item starts unrestricted)
-    EXCLUDE_FIELDS = {
-        "id", "sku", "name", "created_at", "updated_at",
-        "woocommerce_product_id", "squarespace_product_id",
-        "legacy_sku", "upc",
-        "average_cost", "last_cost",
-        "gcode_file_path", "image_url",
-        "customer_id",
-        "parent_product_id", "is_template", "variant_metadata",
-    }
-
-    # Clone product fields
-    clone_data = {}
-    for col in Product.__table__.columns:
-        if col.name not in EXCLUDE_FIELDS:
-            clone_data[col.name] = getattr(source, col.name)
-
-    clone_data["sku"] = new_sku_upper
-    clone_data["name"] = new_name_clean
-    clone_data["has_bom"] = False  # Will be set True if BOM is copied
-
-    new_item = Product(**clone_data)
-    db.add(new_item)
-    db.flush()  # Get the new item's ID
-
-    # Copy active BOM if source has one
-    bom_id = None
-    active_bom = (
-        db.query(BOM)
-        .filter(BOM.product_id == source.id, BOM.active.is_(True))
-        .first()
-    )
-
-    if active_bom:
-        new_bom = BOM(
-            product_id=new_item.id,
-            code=f"{new_sku_upper}-BOM"[:50],
-            name=f"BOM for {new_item.name}"[:255],
-            version=1,
-            revision=active_bom.revision,
-            assembly_time_minutes=active_bom.assembly_time_minutes,
-            effective_date=active_bom.effective_date,
-            notes=f"Duplicated from {source.sku}",
-            active=True,
-        )
-        db.add(new_bom)
-        db.flush()
-
-        # Build override lookup: original_component_id -> new_component_id
-        # NOTE: Keyed by component_id, so if the same component appears on
-        # multiple BOM lines, ALL instances get swapped. This is intentional
-        # for color variants (swap every instance of "PLA Red" to "PLA Blue").
-        override_map = {}
-        if bom_line_overrides:
-            for ov in bom_line_overrides:
-                orig_id = ov.get("original_component_id")
-                new_id = ov.get("new_component_id")
-                if orig_id and new_id:
-                    # Validate new component exists
-                    if not db.query(Product).filter(Product.id == new_id).first():
-                        raise HTTPException(
-                            status_code=400,
-                            detail=f"Override component ID {new_id} not found"
-                        )
-                    override_map[orig_id] = new_id
-
-        # Copy lines with overrides
-        source_lines = (
-            db.query(BOMLine)
-            .filter(BOMLine.bom_id == active_bom.id)
-            .order_by(BOMLine.sequence)
-            .all()
-        )
-        for line in source_lines:
-            component_id = override_map.get(line.component_id, line.component_id)
-            new_line = BOMLine(
-                bom_id=new_bom.id,
-                component_id=component_id,
-                quantity=line.quantity,
-                unit=line.unit,
-                sequence=line.sequence,
-                consume_stage=line.consume_stage,
-                is_cost_only=line.is_cost_only,
-                scrap_factor=line.scrap_factor,
-                notes=line.notes,
-            )
-            db.add(new_line)
-
-        db.flush()
-        new_bom.total_cost = recalculate_bom_cost(new_bom, db)
-        new_item.has_bom = True
-        bom_id = new_bom.id
-
-    # Copy active routing if source has one
-    routing_id = None
-    active_routing = (
-        db.query(Routing)
-        .filter(Routing.product_id == source.id, Routing.is_active.is_(True))
-        .first()
-    )
-
-    if active_routing:
-        new_routing = Routing(
-            product_id=new_item.id,
-            code=f"RTG-{new_sku_upper}"[:50],
-            name=f"Routing for {new_item.name}"[:200],
-            is_template=False,
-            version=1,
-            revision="1.0",
-            is_active=True,
-            effective_date=active_routing.effective_date,
-            notes=f"Duplicated from {source.sku}",
-        )
-        db.add(new_routing)
-        db.flush()
-
-        # Copy operations — track old_id -> new_id for predecessor remapping
-        op_id_map: dict[int, int] = {}
-        source_ops = (
-            db.query(RoutingOperation)
-            .options(joinedload(RoutingOperation.materials))
-            .filter(RoutingOperation.routing_id == active_routing.id)
-            .order_by(RoutingOperation.sequence)
-            .all()
-        )
-
-        for op in source_ops:
-            new_op = RoutingOperation(
-                routing_id=new_routing.id,
-                work_center_id=op.work_center_id,
-                sequence=op.sequence,
-                operation_code=op.operation_code,
-                operation_name=op.operation_name,
-                description=op.description,
-                setup_time_minutes=op.setup_time_minutes,
-                run_time_minutes=op.run_time_minutes,
-                wait_time_minutes=op.wait_time_minutes,
-                move_time_minutes=op.move_time_minutes,
-                runtime_source=op.runtime_source,
-                slicer_file_path=op.slicer_file_path,
-                units_per_cycle=op.units_per_cycle,
-                scrap_rate_percent=op.scrap_rate_percent,
-                labor_rate_override=op.labor_rate_override,
-                machine_rate_override=op.machine_rate_override,
-                can_overlap=op.can_overlap,
-                is_active=op.is_active,
-                # predecessor_operation_id set in second pass below
-            )
-            db.add(new_op)
-            db.flush()
-            op_id_map[op.id] = new_op.id
-
-            # Copy operation materials with component overrides
-            for mat in op.materials:
-                component_id = override_map.get(mat.component_id, mat.component_id) if active_bom else mat.component_id
-                new_mat = RoutingOperationMaterial(
-                    routing_operation_id=new_op.id,
-                    component_id=component_id,
-                    quantity=mat.quantity,
-                    quantity_per=mat.quantity_per,
-                    unit=mat.unit,
-                    scrap_factor=mat.scrap_factor,
-                    is_cost_only=mat.is_cost_only,
-                    is_optional=mat.is_optional,
-                    is_variable=mat.is_variable,
-                    notes=mat.notes,
-                )
-                db.add(new_mat)
-
-        # Second pass: remap predecessor_operation_id references
-        for old_op in source_ops:
-            if old_op.predecessor_operation_id and old_op.predecessor_operation_id in op_id_map:
-                new_op_id = op_id_map[old_op.id]
-                db.query(RoutingOperation).filter(
-                    RoutingOperation.id == new_op_id
-                ).update({
-                    "predecessor_operation_id": op_id_map[old_op.predecessor_operation_id]
-                })
-
-        db.flush()
-        # Use service function (does its own eager-loading) to avoid N+1
-        from app.services.routing_service import recalculate_routing_totals
-        recalculate_routing_totals(new_routing, db)
-        routing_id = new_routing.id
-
-    db.commit()
-    db.refresh(new_item)
-
-    # Build summary message
-    parts = [f"Duplicated from {source.sku}"]
-    if active_bom:
-        parts.append(f"with BOM ({len(source_lines)} lines)")
-    if active_routing:
-        parts.append(f"routing ({len(source_ops)} operations)")
-    if not active_bom and not active_routing:
-        parts.append("(no BOM or routing)")
-
-    return {
-        "id": new_item.id,
-        "sku": new_item.sku,
-        "name": new_item.name,
-        "has_bom": new_item.has_bom,
-        "bom_id": bom_id,
-        "routing_id": routing_id,
-        "message": " ".join(parts),
-    }
+from app.services.bom_management_service import (  # noqa: E402, F401
+    recalculate_bom_cost,
+)
+from app.services.item_cost_service import (  # noqa: E402, F401
+    calculate_item_cost,
+    recost_all_items,
+    recost_item,
+)
+from app.services.item_duplicate_service import (  # noqa: E402, F401
+    duplicate_item,
+)
+from app.services.item_import_service import (  # noqa: E402, F401
+    _COST_COLUMNS,
+    _DESCRIPTION_COLUMNS,
+    _NAME_COLUMNS,
+    _PRICE_COLUMNS,
+    _PURCHASE_UOM_COLUMNS,
+    _SKU_COLUMNS,
+    _UNIT_COLUMNS,
+    _get_best_price_from_row,
+    _get_category_id_from_row,
+    _get_csv_column_value,
+    _get_upc_from_row,
+    _normalize_import_item_type,
+    _parse_price,
+    _strip_html,
+    _update_category_from_row,
+    import_items_from_csv,
+)
